backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
textblob
cohere
pandas
numpy